Move objToJsonLd back into @weco/common
