+++ common/utils/json-ld.ts
@@ -1,0 +1,10 @@
+export function objToJsonLd<T>(obj: T, type: string, root = true) {
+  const jsonObj = JSON.parse(JSON.stringify(obj));
+  const jsonLdAddition = root
+    ? {
+        '@context': 'http://schema.org',
+        '@type': type,
+      }
+    : { '@type': type };
+  return { ...jsonObj, ...jsonLdAddition };
+}
